+++ src/main/java/com/datn/laptopshop/dto/OrderDetailDto.java
@@ -11,15 +11,15 @@ import lombok.Setter;
 @AllArgsConstructor
 @NoArgsConstructor
 public class OrderDetailDto {
-    private Long id;
-    private Long order;
+    private int id;
+    private int order;
     private ProductDto product;
     private int price;
     private int discount;
     private int num;
     private int totalPrice;
 
-    public OrderDetailDto(Long id, Long orderId, int price, int discount, int num, int totalPrice) {
+    public OrderDetailDto(int id, int orderId, int price, int discount, int num, int totalPrice) {
         this.id=id;
         this.order=orderId;
         this.price=price;
